Separate Akka actor support from play26 module
