+++ play-zipkin-tracing/play26/src/main/scala/jp/co/bizreach/trace/play26/implicits/ZipkinTraceImplicits.scala
@@ -1,7 +1,7 @@
 package jp.co.bizreach.trace.play26.implicits
 
 import brave.Span
-import jp.co.bizreach.trace.play26.AkkaSupport.ActorTraceData
+import jp.co.bizreach.trace.akka.actor.ActorTraceSupport.ActorTraceData
 import jp.co.bizreach.trace.{TraceData, ZipkinTraceServiceLike}
 import play.api.mvc.RequestHeader
 
@@ -22,6 +22,12 @@ trait ZipkinTraceImplicits {
     )
   }
 
+  /**
+   * Creates a trace data including a span from request headers for Akka actor.
+   *
+   * @param req the HTTP request header
+   * @return the trace data
+   */
   implicit def request2actorTrace(implicit req: RequestHeader): ActorTraceData = {
     val span = tracer.toSpan(req.headers)((headers, key) => headers.get(key))
     val oneWaySpan = tracer.tracing.tracer.newChild(span.context()).kind(Span.Kind.CLIENT)
