+++ sample/zipkin-api-play26/app/filters/Filters.scala
@@ -1,6 +1,6 @@
 package filters
 
-import javax.inject.{Inject, Named}
+import javax.inject.Inject
 
 import jp.co.bizreach.trace.play26.filter.ZipkinTraceFilter
 import play.api.http.DefaultHttpFilters
